A search mechanism
Use the prompt method to open a dialog box and request the name of a student. When a student name is typed into the box, the program would search through the student records and look for the name that matches. If there's a match, then print out the student related data. And the prompt appears over and over again until the word 'quit' is entered into the prompt dialog box, the program ends.